feature/study: 보고서 관련 임시 로직

+++ src/main/java/com/seoultech/ecc/study/repository/ReviewRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ReviewRepository extends JpaRepository<ReviewEntity, Long> {
-    List<ReviewEntity> findAllByMember_Uuid(Long memberId);
+    List<ReviewEntity> findAllByMember_Uuid(Integer memberId);
 
-    List<ReviewEntity> findAllByReport_ReportIdAndMember_Uuid(Long reportId, Long memberId);
+    List<ReviewEntity> findAllByReport_ReportIdAndMember_Uuid(Long reportId, Integer memberId);
 
     List<ReviewEntity> findAllByReport_ReportId(Long reportId);
 }
